refactor: Use styled.js to import styles
